Moved lib/permutations to lib/arrays
