+++ src/lib/arrays/Permutations.svelte
