+++ src/lib/arrays/permutation.ts
